services/tts/requirements.txt
av==14.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
coloredlogs==15.0.1
ctranslate2==4.5.0
faster-whisper==1.1.1
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.3.0
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
nats-py==2.10.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
onnxruntime==1.19.2
openai-whisper==20240930
packaging==24.2
protobuf==6.30.2
PyAudio==0.2.14
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
sympy==1.13.1
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.13.0
urllib3==2.3.0
whisper==1.1.10
